Begin implementing command manager
## Motivation

We need a component responsible for building and maintaining the current state of the command queue during database operations.

## Solution

Began implementing the command manager but realized we need the journal reader lifecycle component to wait for all of its children to complete their start up processes before the journal reader starts reading and emitting journal entries. In the lifecycle manager,  implemented the ability for lifecycle components to register child lifecycle components and manage the startup of their children. Then updated the journal reader to use this a new feature.

Also added a @db/disk workspace package. Eventually, this will become the injection point for pseudo-random failure injection during simulation testing